refactor(rdp): improve file boundaries for readability
Extract runConnectorForSession() from session.go into connector.go and
move all detection logic (RunStickyKeysCheck, RunUtilmanCheck, detection
sequences, banner formatters) from rdp.go into detect.go.
