+++ internal/plugins/rdp/rdp.go
@@ -21,7 +21,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"net"
-	"os"
 	"strings"
 	"time"
 
@@ -374,254 +373,3 @@ func parseDomainUsername(username string) (domain, user string) {
 }
 
 var classifyError = brutus.NewClassifier(rdpAuthIndicators)
-
-// RunStickyKeysCheck performs sticky keys detection on a separate connection.
-// The noVision flag disables Vision API confirmation.
-func (p *Plugin) RunStickyKeysCheck(ctx context.Context, target string, timeout time.Duration, noVision bool) *StickyKeysResult {
-	host, port := brutus.ParseTarget(target, "3389")
-	addr := net.JoinHostPort(host, port)
-
-	eng, err := initEngine()
-	if err != nil {
-		return &StickyKeysResult{Performed: false, SkipReason: fmt.Sprintf("wasm init: %v", err)}
-	}
-
-	dialer := &net.Dialer{Timeout: timeout}
-	conn, err := dialer.DialContext(ctx, "tcp", addr)
-	if err != nil {
-		return &StickyKeysResult{Performed: false, SkipReason: fmt.Sprintf("connection failed: %v", err)}
-	}
-	defer func() { _ = conn.Close() }()
-
-	inst, err := newInstance(ctx, eng, conn)
-	if err != nil {
-		return &StickyKeysResult{Performed: false, SkipReason: fmt.Sprintf("wasm instance: %v", err)}
-	}
-	defer func() { _ = inst.close(ctx) }()
-
-	stickyResult, err := p.runStickyKeysDetection(ctx, inst, addr, noVision)
-	if err != nil {
-		return &StickyKeysResult{Performed: false, SkipReason: fmt.Sprintf("detection failed: %v", err)}
-	}
-
-	return stickyResult
-}
-
-// runStickyKeysDetection performs the full detection sequence on a non-NLA connection.
-func (p *Plugin) runStickyKeysDetection(ctx context.Context, inst *wasmInstance, addr string, noVision bool) (*StickyKeysResult, error) {
-	result := &StickyKeysResult{Performed: true}
-
-	cfg := rdpConfig{
-		Server:   addr,
-		Username: "",
-		Password: "",
-		Domain:   "",
-		SkipAuth: true,
-	}
-	configBytes, err := json.Marshal(cfg)
-	if err != nil {
-		return nil, fmt.Errorf("marshal config: %w", err)
-	}
-
-	connHandle, _, err := p.runConnectorForSession(ctx, inst, configBytes)
-	if err != nil {
-		result.Performed = false
-		result.SkipReason = fmt.Sprintf("connection failed: %v", err)
-		return result, nil
-	}
-	// Ensure connector handle is freed after session use
-	callCtx := inst.callCtx(ctx)
-	defer func() {
-		if freeFn := inst.mod.ExportedFunction("connector_free"); freeFn != nil {
-			_, _ = freeFn.Call(callCtx, uint64(connHandle))
-		}
-	}()
-
-	baseline, response, width, height, err := p.runSession(ctx, inst, connHandle, 1024, 768)
-	if err != nil {
-		result.Performed = false
-		result.SkipReason = fmt.Sprintf("session failed: %v", err)
-		return result, nil
-	}
-
-	// Vision API confirmation is optional: requires ANTHROPIC_API_KEY and
-	// can be disabled with --no-vision flag.
-	var visionAPIKey string
-	if !noVision {
-		visionAPIKey = os.Getenv("ANTHROPIC_API_KEY")
-	}
-	*result = runStickyKeysAnalysis(ctx, baseline, response, width, height, visionAPIKey)
-	result.Performed = true
-
-	return result, nil
-}
-
-// RunUtilmanCheck performs utilman backdoor detection on a separate connection.
-func (p *Plugin) RunUtilmanCheck(ctx context.Context, target string, timeout time.Duration, noVision bool) *UtilmanResult {
-	host, port := brutus.ParseTarget(target, "3389")
-	addr := net.JoinHostPort(host, port)
-
-	eng, err := initEngine()
-	if err != nil {
-		return &UtilmanResult{Performed: false, SkipReason: fmt.Sprintf("wasm init: %v", err)}
-	}
-
-	dialer := &net.Dialer{Timeout: timeout}
-	conn, err := dialer.DialContext(ctx, "tcp", addr)
-	if err != nil {
-		return &UtilmanResult{Performed: false, SkipReason: fmt.Sprintf("connection failed: %v", err)}
-	}
-	defer func() { _ = conn.Close() }()
-
-	inst, err := newInstance(ctx, eng, conn)
-	if err != nil {
-		return &UtilmanResult{Performed: false, SkipReason: fmt.Sprintf("wasm instance: %v", err)}
-	}
-	defer func() { _ = inst.close(ctx) }()
-
-	utilmanResult, err := p.runUtilmanDetection(ctx, inst, addr, noVision)
-	if err != nil {
-		return &UtilmanResult{Performed: false, SkipReason: fmt.Sprintf("detection failed: %v", err)}
-	}
-
-	return utilmanResult
-}
-
-// runUtilmanDetection performs the full utilman detection sequence on a non-NLA connection.
-func (p *Plugin) runUtilmanDetection(ctx context.Context, inst *wasmInstance, addr string, noVision bool) (*UtilmanResult, error) {
-	result := &UtilmanResult{Performed: true}
-
-	cfg := rdpConfig{
-		Server:   addr,
-		Username: "",
-		Password: "",
-		Domain:   "",
-		SkipAuth: true,
-	}
-	configBytes, err := json.Marshal(cfg)
-	if err != nil {
-		return nil, fmt.Errorf("marshal config: %w", err)
-	}
-
-	connHandle, _, err := p.runConnectorForSession(ctx, inst, configBytes)
-	if err != nil {
-		result.Performed = false
-		result.SkipReason = fmt.Sprintf("connection failed: %v", err)
-		return result, nil
-	}
-	// Ensure connector handle is freed after session use
-	callCtx := inst.callCtx(ctx)
-	defer func() {
-		if freeFn := inst.mod.ExportedFunction("connector_free"); freeFn != nil {
-			_, _ = freeFn.Call(callCtx, uint64(connHandle))
-		}
-	}()
-
-	baseline, response, width, height, err := p.runUtilmanSession(ctx, inst, connHandle, 1024, 768)
-	if err != nil {
-		result.Performed = false
-		result.SkipReason = fmt.Sprintf("session failed: %v", err)
-		return result, nil
-	}
-
-	// Vision API confirmation is optional: requires ANTHROPIC_API_KEY and
-	// can be disabled with --no-vision flag.
-	var visionAPIKey string
-	if !noVision {
-		visionAPIKey = os.Getenv("ANTHROPIC_API_KEY")
-	}
-	*result = runUtilmanAnalysis(ctx, baseline, response, width, height, visionAPIKey)
-	result.Performed = true
-
-	return result, nil
-}
-
-// formatUtilmanBanner appends utilman detection results to the banner.
-func formatUtilmanBanner(existingBanner string, result *UtilmanResult) string {
-	if result == nil {
-		return existingBanner
-	}
-	if !result.Performed {
-		banner := existingBanner
-		if banner != "" {
-			banner += "\n"
-		}
-		if result.SkipReason != "" {
-			banner += "[INFO] Utilman check skipped: " + result.SkipReason
-		}
-		return banner
-	}
-
-	banner := existingBanner
-	if banner != "" {
-		banner += "\n"
-	}
-
-	switch result.OverallVerdict {
-	case "backdoor_confirmed":
-		banner += fmt.Sprintf("[CRITICAL] Utilman backdoor CONFIRMED (confidence: %.0f%%)\n", result.Confidence*100)
-		banner += "utilman.exe has been replaced with cmd.exe or similar.\n"
-		banner += "SYSTEM-level unauthenticated access available via Win+U on login screen.\n"
-		banner += "B-TP: malicious persistence (T1546.008), forgotten password recovery, or pentest artifact.\n"
-		banner += "Remediation: Boot from Windows install media, restore original utilman.exe, or run sfc /scannow."
-	case "backdoor_likely":
-		banner += fmt.Sprintf("[HIGH] Utilman backdoor likely (confidence: %.0f%%)\n", result.Confidence*100)
-		banner += "A window appeared after Win+U on the login screen.\n"
-		banner += "Heuristic: " + result.HeuristicResult
-		if result.VisionResult != "" {
-			banner += " | Vision: " + result.VisionResult
-		}
-	case "vulnerable":
-		banner += "[INFO] Non-NLA RDP target. Utilman triggers normally (no backdoor detected).\n"
-		banner += "Target is vulnerable if utilman.exe is later replaced."
-	case "clean":
-		banner += "[INFO] Utilman check: clean (no response to Win+U)."
-	}
-
-	return banner
-}
-
-// formatStickyKeysBanner appends sticky keys detection results to the banner.
-func formatStickyKeysBanner(existingBanner string, result *StickyKeysResult) string {
-	if result == nil {
-		return existingBanner
-	}
-	if !result.Performed {
-		banner := existingBanner
-		if banner != "" {
-			banner += "\n"
-		}
-		if result.SkipReason != "" {
-			banner += "[INFO] Sticky keys check skipped: " + result.SkipReason
-		}
-		return banner
-	}
-
-	banner := existingBanner
-	if banner != "" {
-		banner += "\n"
-	}
-
-	switch result.OverallVerdict {
-	case "backdoor_confirmed":
-		banner += fmt.Sprintf("[CRITICAL] Sticky keys backdoor CONFIRMED (confidence: %.0f%%)\n", result.Confidence*100)
-		banner += "sethc.exe has been replaced with cmd.exe or similar.\n"
-		banner += "SYSTEM-level unauthenticated access available via 5x Shift.\n"
-		banner += "B-TP: malicious persistence (T1546.008), forgotten password recovery, or pentest artifact.\n"
-		banner += "Remediation: Boot from Windows install media, restore original sethc.exe, or run sfc /scannow."
-	case "backdoor_likely":
-		banner += fmt.Sprintf("[HIGH] Sticky keys backdoor likely (confidence: %.0f%%)\n", result.Confidence*100)
-		banner += "A dark window appeared after 5x Shift on the login screen.\n"
-		banner += "Heuristic: " + result.HeuristicResult
-		if result.VisionResult != "" {
-			banner += " | Vision: " + result.VisionResult
-		}
-	case "vulnerable":
-		banner += "[INFO] Non-NLA RDP target. Sticky Keys triggers normally (no backdoor detected).\n"
-		banner += "Target is vulnerable if sethc.exe is later replaced."
-	case "clean":
-		banner += "[INFO] Sticky keys check: clean (no response to 5x Shift)."
-	}
-
-	return banner
-}
